removed old stuff, functional new UI
