--- src/main/java/com/flansmod/warforge/client/ModularUISiegeCamp.java
@@ -1,4 +1,0 @@
-package com.flansmod.warforge.client;
-
-public class ModularUISiegeCamp {
-}
